rename plugin to plugin.js
